+++ src/main/java/com/kruemblegard/client/render/model/MoogloomModel.java
@@ -1,0 +1,26 @@
+package com.kruemblegard.client.render.model;
+
+import com.kruemblegard.Kruemblegard;
+import com.kruemblegard.entity.MoogloomEntity;
+
+import net.minecraft.resources.ResourceLocation;
+
+import software.bernie.geckolib.model.GeoModel;
+
+public class MoogloomModel extends GeoModel<MoogloomEntity> {
+
+    @Override
+    public ResourceLocation getModelResource(MoogloomEntity animatable) {
+        return new ResourceLocation(Kruemblegard.MOD_ID, "geo/moogloom.geo.json");
+    }
+
+    @Override
+    public ResourceLocation getTextureResource(MoogloomEntity animatable) {
+        return new ResourceLocation(Kruemblegard.MOD_ID, "textures/entity/moogloom.png");
+    }
+
+    @Override
+    public ResourceLocation getAnimationResource(MoogloomEntity animatable) {
+        return new ResourceLocation(Kruemblegard.MOD_ID, "animations/moogloom.animation.json");
+    }
+}