+++ src/main/java/com/kruemblegard/client/render/model/TraderBeetleModel.java
@@ -1,7 +1,7 @@
 package com.kruemblegard.client.render.model;
 
 import com.kruemblegard.Kruemblegard;
-import com.kruemblegard.entity.ScaralonBeetleEntity;
+import com.kruemblegard.entity.TraderBeetleEntity;
 
 import net.minecraft.core.Direction;
 import net.minecraft.resources.ResourceLocation;
@@ -9,17 +9,19 @@ import net.minecraft.util.Mth;
 
 import software.bernie.geckolib.core.animatable.model.CoreGeoBone;
 import software.bernie.geckolib.core.animation.AnimationState;
-
 import software.bernie.geckolib.model.GeoModel;
 
-public class ScaralonBeetleModel extends GeoModel<ScaralonBeetleEntity> {
+/**
+ * Uses the same Geo/animations as Scaralon beetles, but swaps to trader-beetle textures.
+ */
+public class TraderBeetleModel extends GeoModel<TraderBeetleEntity> {
     private static final String SADDLE_BONE = "saddle";
     private static final String CHEST_BONE = "chest";
     private static final String CARPET_BONE = "carpet";
     private static final String LARVA_ROOT_BONE = "root";
 
     @Override
-    public ResourceLocation getModelResource(ScaralonBeetleEntity animatable) {
+    public ResourceLocation getModelResource(TraderBeetleEntity animatable) {
         if (animatable.isBaby()) {
             return new ResourceLocation(Kruemblegard.MOD_ID, "geo/scaralon_larva.geo.json");
         }
@@ -27,7 +29,7 @@ public class ScaralonBeetleModel extends GeoModel<ScaralonBeetleEntity> {
     }
 
     @Override
-    public ResourceLocation getTextureResource(ScaralonBeetleEntity animatable) {
+    public ResourceLocation getTextureResource(TraderBeetleEntity animatable) {
         if (animatable.isBaby()) {
             return new ResourceLocation(Kruemblegard.MOD_ID, "textures/entity/scaralon_beetle/scaralon_larva.png");
         }
@@ -36,7 +38,7 @@ public class ScaralonBeetleModel extends GeoModel<ScaralonBeetleEntity> {
     }
 
     @Override
-    public ResourceLocation getAnimationResource(ScaralonBeetleEntity animatable) {
+    public ResourceLocation getAnimationResource(TraderBeetleEntity animatable) {
         if (animatable.isBaby()) {
             return new ResourceLocation(Kruemblegard.MOD_ID, "animations/scaralon_larva.animation.json");
         }
@@ -44,11 +46,10 @@ public class ScaralonBeetleModel extends GeoModel<ScaralonBeetleEntity> {
     }
 
     @Override
-    public void setCustomAnimations(ScaralonBeetleEntity animatable, long instanceId, AnimationState<ScaralonBeetleEntity> animationState) {
+    public void setCustomAnimations(TraderBeetleEntity animatable, long instanceId, AnimationState<TraderBeetleEntity> animationState) {
         super.setCustomAnimations(animatable, instanceId, animationState);
 
         if (animatable.isBaby()) {
-            // Larva wall-climb alignment: rotate the larva onto the surface it's attached to.
             CoreGeoBone root = this.getAnimationProcessor().getBone(LARVA_ROOT_BONE);
             if (root != null && animatable.isLarvaClimbing()) {
                 Direction face = animatable.getLarvaClimbFace();
@@ -57,8 +58,6 @@ public class ScaralonBeetleModel extends GeoModel<ScaralonBeetleEntity> {
                 float rotZ = 0.0F;
 
                 if (face != null) {
-                    // Align larva "belly" (down) to the wall normal.
-                    // Ground: no override; Walls: rotate 90 degrees towards the face.
                     switch (face) {
                         case NORTH -> rotX = 90.0F * Mth.DEG_TO_RAD;
                         case SOUTH -> rotX = -90.0F * Mth.DEG_TO_RAD;
@@ -76,26 +75,22 @@ public class ScaralonBeetleModel extends GeoModel<ScaralonBeetleEntity> {
             return;
         }
 
-        // Hide saddle geometry unless the beetle is actually saddled.
         CoreGeoBone saddle = this.getAnimationProcessor().getBone(SADDLE_BONE);
         if (saddle != null) {
             saddle.setHidden(!animatable.isSaddled());
         }
 
-        // Hide chest geometry unless the beetle actually has a chest attached.
         CoreGeoBone chest = this.getAnimationProcessor().getBone(CHEST_BONE);
         if (chest != null) {
             chest.setHidden(!animatable.hasChest());
         }
 
-        // Hide carpet geometry unless decor is applied.
-        boolean hasCarpet = animatable.hasCarpet();
         CoreGeoBone carpet = this.getAnimationProcessor().getBone(CARPET_BONE);
         if (carpet != null) {
-            carpet.setHidden(!hasCarpet);
+            carpet.setHidden(!animatable.hasCarpet());
         }
 
-        // If the model has split carpet bones, keep them consistent too.
+        boolean hasCarpet = animatable.hasCarpet();
         CoreGeoBone carpetFrame = this.getAnimationProcessor().getBone("carpet_frame");
         if (carpetFrame != null) {
             carpetFrame.setHidden(!hasCarpet);