+++ src/main/java/com/kruemblegard/registry/ModEntities.java
@@ -8,6 +8,7 @@ import com.kruemblegard.entity.MoogloomEntity;
 import com.kruemblegard.entity.PebblitEntity;
 import com.kruemblegard.entity.ScaralonBeetleEntity;
 import com.kruemblegard.entity.ScatteredEndermanEntity;
+import com.kruemblegard.entity.TraderBeetleEntity;
 import com.kruemblegard.entity.TraprockEntity;
 import com.kruemblegard.entity.WyrdwingEntity;
 import com.kruemblegard.entity.vehicle.KruemblegardBoatEntity;
@@ -86,6 +87,15 @@ public class ModEntities {
                     ScaralonBeetleEntity::new, MobCategory.CREATURE)
                 .sized(1.2f, 1.4f)
                 .build(new ResourceLocation(Kruemblegard.MOD_ID, "scaralon_beetle").toString())
+        );
+
+    public static final RegistryObject<EntityType<TraderBeetleEntity>> TRADER_BEETLE =
+        ENTITIES.register(
+            "trader_beetle",
+            () -> EntityType.Builder.<TraderBeetleEntity>of(
+                    TraderBeetleEntity::new, MobCategory.CREATURE)
+                .sized(1.2f, 1.4f)
+                .build(new ResourceLocation(Kruemblegard.MOD_ID, "trader_beetle").toString())
         );
 
     public static final RegistryObject<EntityType<WyrdwingEntity>> WYRDWING =